Update GenerateIndex.js
Some users are experiencing issues with row.title is passed from DB to client with title as type Integer. This can break index exports, as a.title.toLowerCase(), where a is an integer throws "function does not exist" errors, thus breaking the export of the index. Casting data.title to type string, appears to fix this issue. 

Currently, I see no consequences in casting each data type as string, as a safeguard moving forward.